[refactor] habit 데이터와 다루는 함수를 Page  컴포넌트에 옮김
