+++ src/component/App.jsx
@@ -3,31 +3,9 @@ import Page from './Page.jsx';
 
 
 function App() {
-  const [state, setState] = useState({
-
-    habits: []
-  })
-  const { habits } = state;
-
-  function handleDelete(key) {
-    setState({
-      habits: habits.filter((value, index) => index != key)
-    }
-    )
-  }
-
-  function handleSubmit(e) {
-    e.preventDefault()
-    setState({
-      ...state,
-      habits: [e.target[0].value, ...habits]
-    })
-    e.target[0].value = "";
-  }
-
 
   return (
-    <Page onSubmit={handleSubmit} onClickDelete={handleDelete} habits={habits} />
+    <Page />
   )
 }
