+++ src/component/Page.jsx
@@ -1,8 +1,29 @@
-import React from 'react'
+import React, { useState } from 'react'
 import Input from './Input.jsx'
-function Page({ onSubmit, onClickDelete, habits }) {
+function Page() {
+
+  const [state, setState] = useState({
+    habits: []
+  })
+  const { habits } = state;
+
+  function handleDelete(key) {
+    setState({
+      habits: habits.filter((value, index) => index != key)
+    }
+    )
+  }
+
+  function handleSubmit(e) {
+    e.preventDefault()
+    setState({
+      habits: [e.target[0].value, ...habits]
+    })
+    e.target[0].value = "";
+  }
+
   return (
-    <Input onSubmit={onSubmit} onClickDelete={onClickDelete} habits={habits} />
+    <Input onSubmit={handleSubmit} onClickDelete={handleDelete} habits={habits} />
   )
 }
